requirements.txt
paho-mqtt==1.4.0
slackclient==1.3.1
